+++ clang/test/CXX/drs/dr14xx.cpp
@@ -7,6 +7,8 @@
 // expected-no-diagnostics
 #endif
 
+// dr1425: na abi
+
 namespace dr1460 { // dr1460: 3.5
 #if __cplusplus >= 201103L
   namespace DRExample {